Updated version
Updated version contains all the visualizations which can be shown to insurance companies for basic idea of how many of their customers are at high risk. Or which policy to offer to customers of specific risk level. It is not integrated due to short of time but still we have attached the images and html file in templates.